add/edit post form, post & profile pages and ui changes

+++ src/server/trpc/router/post.ts
@@ -1,3 +1,4 @@
+import { TRPCError } from "@trpc/server";
 import { z } from "zod";
 import { postSchemaInput } from "../../../schemas/post.schema";
 import { protectedProcedure, publicProcedure, router } from "../trpc";
@@ -6,10 +7,21 @@ export const postRouter = router({
   getAll: publicProcedure.query(async ({ ctx }) => {
     return await ctx.prisma.post.findMany({
       orderBy: {
-        createdAt: "desc",
+        updatedAt: "desc",
       },
     });
   }),
+  getById: protectedProcedure
+    .input(z.object({ postId: z.string() }))
+    .query(async ({ ctx, input }) => {
+      const { postId } = input;
+      const foundPost = await ctx.prisma.post.findUnique({
+        where: {
+          id: postId,
+        },
+      });
+      return foundPost;
+    }),
   getUsersPosts: protectedProcedure
     .input(z.object({ userId: z.string() }))
     .query(async ({ ctx, input }) => {
@@ -41,6 +53,7 @@ export const postRouter = router({
           creatorId: user.id,
           creatorName: user.name,
           creatorImage: user.image,
+          updatedAt: undefined,
         },
       });
 
@@ -50,6 +63,19 @@ export const postRouter = router({
     .input(z.object({ postId: z.string() }))
     .mutation(async ({ ctx, input }) => {
       const { postId } = input;
+
+      const user = await prisma?.user.findUnique({
+        where: { id: ctx.session.user.id },
+        select: { posts: true },
+      });
+
+      if (!user?.posts.find((post) => post.id === postId)) {
+        throw new TRPCError({
+          code: "UNAUTHORIZED",
+          message: "You cannot remove not your own posts!",
+        });
+      }
+
       const deletedPost = await ctx.prisma.post.delete({
         where: {
           id: postId,
@@ -58,9 +84,23 @@ export const postRouter = router({
       return deletedPost;
     }),
   update: protectedProcedure
-    .input(postSchemaInput)
+    .input(z.object({ postId: z.string(), postSchemaInput }))
     .mutation(async ({ ctx, input }) => {
-      const { title, description, sport, workoutDate, id: postId } = input;
+      const user = await prisma?.user.findUnique({
+        where: { id: ctx.session.user.id },
+        select: { posts: true },
+      });
+
+      const { postId } = input;
+
+      if (!user?.posts.find((post) => post.id === postId)) {
+        throw new TRPCError({
+          code: "UNAUTHORIZED",
+          message: "You cannot edit not your own posts!",
+        });
+      }
+
+      const { title, description, sport, workoutDate } = input.postSchemaInput;
       const workoutDateTime = new Date(workoutDate);
 
       const updatePost = await ctx.prisma.post.update({
@@ -70,6 +110,7 @@ export const postRouter = router({
           description,
           sport,
           workoutDate: workoutDateTime,
+          updatedAt: new Date(),
         },
       });
